Move test fixtures and integration tests
