+++ src/test/with-features.js
@@ -4,7 +4,7 @@ import dom from 'cheerio';
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import createWrappedComponent from '../test-fixtures/create-wrapped-component';
-import createFeature from './fixtures/createFeature';
+import createFeature from '../test-fixtures/create-feature';
 
 const render = ReactDOMServer.renderToStaticMarkup;
